# Optimize query to check that plant_id exists based on plant properties